Affix the command and graphics column on sized desktop views and larger.
Hid the commands view in sized tablet views and below. To add the off canvas commands slide out.
Adjusted the height of the drop container when dragging to allow a large dragging item to be droppable at the bottom of the drop container.
Restyled the navigation.
Refactored the main javascript file.