Implemented variable manager to manage all the user defined variables.
Populated the variable manager view for the user to manage the created variables.
Synchronized all the variable combo boxes in the commands to sync with the variable table.
Synchronized the variable manage view with the variable table.
Styled the variable table and its entries.
Fixed affix width when page scroll.